Update_1701
Dodano 5 prefabów, dodano ich kolorowe wersje, dodano przyciski, zmiana położenia UI, zmiana prefabów, dodano materiały i meshe, zmiana w skrypcie:
dodano building postition, zmiana w building placement( teraz możliwe są większe obiekty)

+++ Assets/Scripts/MapModule/MapData.cs
@@ -7,8 +7,9 @@ using UnityEngine;
 public class MapData
 {
     public string mapName;
-    public double budget;
+    public int budget;
     public int columnsNumber;
     public int rowsNumber;
     public Tile[] tiles;
+    //public MapHeat mapHeat;
 }
